requirements.txt
torch
torchvision
Pillow
numpy
packaging
color-matcher